requirements.txt
asgiref==3.3.1
certifi==2019.9.11
Django==3.1.5
djangorestframework==3.12.2
Pillow==8.1.0
pipenv==2018.11.26
pkg-resources==0.0.0
pytz==2020.5
sqlparse==0.4.1
virtualenv==16.7.7
virtualenv-clone==0.5.3
